+++ cnccode/business_classes/BURenQuotation.inc.php
@@ -31,7 +31,7 @@ class BURenQuotation extends Business
     function updateRenQuotation(&$dsData)
     {
         $this->setMethodName('updateRenQuotation');
-        $this->updateDataaccessObject($dsData, $this->dbeRenQuotation);
+        $this->updateDataAccessObject($dsData, $this->dbeRenQuotation);
 
         return TRUE;
     }
@@ -255,7 +255,12 @@ class BURenQuotation extends Business
         $previousCustomerID = 99999;
 
         while ($this->dbeJRenQuotation->fetchNext()) {
-
+            ?>
+            quotation
+            <div>
+                contract number: <?= $dbeJCustomerItem->getValue(DBECustomerItem::customerItemID) ?>
+            </div>
+            <?php
             if ($dbeJCustomerItem->getRow($this->dbeJRenQuotation->getValue('customerItemID'))) {
                 /*
                  * Group many renewals for same customer under one quote