+++ cnccode/data_classes/DBEJRenDomain.inc.php
@@ -7,35 +7,90 @@ require_once($cfg["path_dbe"] . "/DBECustomerItem.inc.php");
 
 class DBEJRenDomain extends DBECustomerItem
 {
+
+
     function __construct(&$owner)
     {
         parent::__construct($owner);
         $this->setAddColumnsOn();
-        $this->addColumn("customerName", DA_STRING, DA_NOT_NULL, "cus_name");
-        $this->addColumn("siteName", DA_STRING, DA_NOT_NULL, "CONCAT(add_add1, ' ', add_town, ' ' , add_postcode)");
-        $this->addColumn("itemDescription", DA_STRING, DA_NOT_NULL, "itm_desc");
-        $this->addColumn("itemTypeDescription", DA_STRING, DA_NOT_NULL, "ity_desc");
-        $this->addColumn("itemID", DA_ID, DA_NOT_NULL, "itm_itemno");
-        $this->addColumn("invoiceFromDate", DA_DATE, DA_NOT_NULL,
-            "DATE_FORMAT( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` MONTH ), '%d/%m/%Y')");
-        $this->addColumn("invoiceToDate", DA_DATE, DA_NOT_NULL,
+        $this->addColumn(
+            "customerName",
+            DA_STRING,
+            DA_NOT_NULL,
+            "cus_name"
+        );
+        $this->addColumn(
+            "siteName",
+            DA_STRING,
+            DA_NOT_NULL,
+            "CONCAT(add_add1, ' ', add_town, ' ' , add_postcode)"
+        );
+        $this->addColumn(
+            "itemDescription",
+            DA_STRING,
+            DA_NOT_NULL,
+            "itm_desc"
+        );
+        $this->addColumn(
+            "itemTypeDescription",
+            DA_STRING,
+            DA_NOT_NULL,
+            "ity_desc"
+        );
+        $this->addColumn(
+            "itemID",
+            DA_ID,
+            DA_NOT_NULL,
+            "itm_itemno"
+        );
+        $this->addColumn(
+            "invoiceFromDate",
+            DA_DATE,
+            DA_NOT_NULL,
+            "DATE_FORMAT( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` MONTH ), '%d/%m/%Y')"
+        );
+        $this->addColumn(
+            "invoiceToDate",
+            DA_DATE,
+            DA_NOT_NULL,
             "DATE_FORMAT(
  				DATE_SUB(
  					DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` + `invoicePeriodMonths` MONTH ),
  					INTERVAL 1 DAY
  				)
- 				, '%d/%m/%Y')");
-        $this->addColumn("invoiceFromDateYMD", DA_DATE, DA_NOT_NULL,
-            "DATE_FORMAT( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` MONTH ), '%%Y-%m-%d') as invoiceFromDateYMD");
-        $this->addColumn("invoiceToDateYMD", DA_DATE, DA_NOT_NULL,
+ 				, '%d/%m/%Y')"
+        );
+        $this->addColumn(
+            "invoiceFromDateYMD",
+            DA_DATE,
+            DA_NOT_NULL,
+            "DATE_FORMAT( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` MONTH ), '%%Y-%m-%d') as invoiceFromDateYMD"
+        );
+        $this->addColumn(
+            "invoiceToDateYMD",
+            DA_DATE,
+            DA_NOT_NULL,
             "DATE_FORMAT(
          DATE_SUB(
            DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` + `invoicePeriodMonths` MONTH ),
            INTERVAL 1 DAY
          )
-         , '%Y-%m-%d')as invoiceToDateYMD");
-        $this->addColumn("salePrice", DA_FLOAT, DA_NOT_NULL, "itm_sstk_price");
-        $this->addColumn("costPrice", DA_FLOAT, DA_NOT_NULL, "itm_sstk_cost");
+         , '%Y-%m-%d')as invoiceToDateYMD"
+        );
+        $this->addColumn(
+            "salePrice",
+            DA_FLOAT,
+            DA_NOT_NULL,
+            "itm_sstk_price"
+        );
+        $this->addColumn(
+            "costPrice",
+            DA_FLOAT,
+            DA_NOT_NULL,
+            "itm_sstk_cost"
+        );
+
+
         $this->setAddColumnsOff();
     }
 
@@ -117,10 +172,9 @@ class DBEJRenDomain extends DBECustomerItem
 				DATE_FORMAT( DATE_ADD( CURDATE(), INTERVAL 1 MONTH ), '%Y%m' ) >=
 				DATE_FORMAT( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` MONTH ), '%Y%m' )
 			 	AND declinedFlag = 'N'
-        AND renewalTypeID = 4
-			 ORDER BY cui_custno
+        AND renewalTypeID = 4 and directDebitFlag <> 'Y'
+			 ORDER BY cui_custno, autoGenerateContractInvoice asc
 		";
-
         $this->setQueryString($statement);
         $ret = (parent::getRows());
     }
@@ -139,9 +193,12 @@ class DBEJRenDomain extends DBECustomerItem
       JOIN itemtype ON  ity_itemtypeno = itm_itemtypeno
       JOIN customer ON  cus_custno = cui_custno
       JOIN address ON  add_custno = cui_custno AND add_siteno = cui_siteno
-     WHERE cui_cuino ID IN ('" . implode('\',\'', $ids) . "')" .
+     WHERE cui_cuino ID IN ('" . implode(
+                '\',\'',
+                $ids
+            ) . "')" .
             " AND declinedFlag = 'N'
-      AND renewalTypeID = 4
+      AND renewalTypeID = 4 and directDebitFlag <> 'Y'
      ORDER BY cui_custno";
 
         $this->setQueryString($statement);