+++ cnccode/data_classes/DBEJRenHosting.inc.php
@@ -7,29 +7,101 @@ require_once($cfg["path_dbe"] . "/DBECustomerItem.inc.php");
 
 class DBEJRenHosting extends DBECustomerItem
 {
+    const customerName = "customerName";
+    const siteName = "siteName";
+    const itemDescription = "itemDescription";
+    const itemTypeDescription = "itemTypeDescription";
+    const itemID = "itemID";
+    const invoiceFromDate = "invoiceFromDate";
+    const invoiceToDate = "invoiceToDate";
+    const invoiceFromDateYMD = "invoiceFromDateYMD";
+    const invoiceToDateYMD = "invoiceToDateYMD";
+    const curUnitSale = "curUnitSale";
+    const curUnitCost = "curUnitCost";
+    const allowDirectDebit = "allowDirectDebit";
+
     function __construct(&$owner)
     {
         parent::__construct($owner);
         $this->setAddColumnsOn();
-        $this->addColumn("customerName", DA_STRING, DA_NOT_NULL, "cus_name");
-        $this->addColumn("siteName", DA_STRING, DA_NOT_NULL, "CONCAT(add_add1, ' ', add_town, ' ' , add_postcode)");
-        $this->addColumn("itemDescription", DA_STRING, DA_NOT_NULL, "itm_desc");
-        $this->addColumn("itemTypeDescription", DA_STRING, DA_NOT_NULL, "ity_desc");
-        $this->addColumn("itemID", DA_ID, DA_NOT_NULL, "itm_itemno");
-        $this->addColumn("invoiceFromDate", DA_DATE, DA_NOT_NULL,
-            "DATE_FORMAT( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` MONTH ), '%d/%m/%Y')");
-        $this->addColumn("invoiceToDate", DA_DATE, DA_NOT_NULL,
+        $this->addColumn(
+            "customerName",
+            DA_STRING,
+            DA_NOT_NULL,
+            "cus_name"
+        );
+        $this->addColumn(
+            "siteName",
+            DA_STRING,
+            DA_NOT_NULL,
+            "CONCAT(add_add1, ' ', add_town, ' ' , add_postcode)"
+        );
+        $this->addColumn(
+            "itemDescription",
+            DA_STRING,
+            DA_NOT_NULL,
+            "itm_desc"
+        );
+        $this->addColumn(
+            "itemTypeDescription",
+            DA_STRING,
+            DA_NOT_NULL,
+            "ity_desc"
+        );
+        $this->addColumn(
+            "itemID",
+            DA_ID,
+            DA_NOT_NULL,
+            "itm_itemno"
+        );
+        $this->addColumn(
+            "invoiceFromDate",
+            DA_DATE,
+            DA_NOT_NULL,
+            "DATE_FORMAT( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` MONTH ), '%d/%m/%Y')"
+        );
+        $this->addColumn(
+            "invoiceToDate",
+            DA_DATE,
+            DA_NOT_NULL,
             "DATE_FORMAT(
  				DATE_SUB(
  					DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` + `invoicePeriodMonths` MONTH ),
  					INTERVAL 1 DAY
  				)
- 				, '%d/%m/%Y')");
-        $this->addColumn("invoiceFromDateYMD", DA_DATE, DA_NOT_NULL,
-            "DATE_FORMAT( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` MONTH ), '%Y-%m-%d') as invoiceFromDateYMD");
-        $this->addColumn("invoiceToDateYMD", DA_DATE, DA_NOT_NULL, "DATE_FORMAT( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` + `invoicePeriodMonths` MONTH ), '%Y-%m-%d') as invoiceToDateYMD");
-        $this->addColumn("curUnitSale", DA_FLOAT, DA_NOT_NULL, 'cui_sale_price');
-        $this->addColumn("curUnitCost", DA_FLOAT, DA_NOT_NULL, 'cui_cost_price');
+ 				, '%d/%m/%Y')"
+        );
+        $this->addColumn(
+            "invoiceFromDateYMD",
+            DA_DATE,
+            DA_NOT_NULL,
+            "DATE_FORMAT( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` MONTH ), '%Y-%m-%d') as invoiceFromDateYMD"
+        );
+        $this->addColumn(
+            "invoiceToDateYMD",
+            DA_DATE,
+            DA_NOT_NULL,
+            "DATE_FORMAT( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` + `invoicePeriodMonths` MONTH ), '%Y-%m-%d') as invoiceToDateYMD"
+        );
+        $this->addColumn(
+            "curUnitSale",
+            DA_FLOAT,
+            DA_NOT_NULL,
+            'cui_sale_price'
+        );
+        $this->addColumn(
+            "curUnitCost",
+            DA_FLOAT,
+            DA_NOT_NULL,
+            'cui_cost_price'
+        );
+
+        $this->addColumn(
+            self::allowDirectDebit,
+            DA_YN,
+            DA_NOT_NULL
+        );
+
         $this->setAddColumnsOff();
     }
 
@@ -98,7 +170,6 @@ class DBEJRenHosting extends DBECustomerItem
      *
      * WHen the invoice has been generated, the total invoice months is increased by the invoice period months
      * so the renewal gets picked up again.
-     *
      */
     function getRenewalsDueRows()
     {
@@ -112,10 +183,8 @@ class DBEJRenHosting extends DBECustomerItem
 			JOIN customer ON  cus_custno = cui_custno
       JOIN address ON  add_custno = cui_custno AND add_siteno = cui_siteno
 		 WHERE CURDATE() >= ( DATE_ADD(`installationDate`, INTERVAL `totalInvoiceMonths` - 1 MONTH ) )
-		 AND declinedFlag = 'N'
-     AND renewalTypeID = 5      
-		 ORDER BY cui_custno
-		";
+		 AND declinedFlag = 'N' AND renewalTypeID = 5 and directDebitFlag <> 'Y'";
+        $statement .= " ORDER BY cui_custno, autoGenerateContractInvoice asc";
 
         $this->setQueryString($statement);
         $ret = (parent::getRows());
@@ -135,9 +204,12 @@ class DBEJRenHosting extends DBECustomerItem
       JOIN itemtype ON  ity_itemtypeno = itm_itemtypeno
       JOIN customer ON  cus_custno = cui_custno
       JOIN address ON  add_custno = cui_custno AND add_siteno = cui_siteno
-     WHERE customerItemID IN ('" . implode('\',\'', $ids) . "')" .
+     WHERE customerItemID IN ('" . implode(
+                '\',\'',
+                $ids
+            ) . "')" .
             " AND declinedFlag = 'N'
-        AND renewalTypeID = 5      
+        AND renewalTypeID = 5  and directDebitFlag <> 'Y'    
       ORDER BY cui_custno
      ";